refactor: Validator의 성격보다 Utils의 성격이 더 적합하다고 판단되는 메서드 이동
메서드 이동에 따른 불필요 코드 및 Test 코드 변경
