+++ src/test/java/calculator/utils/StringUtilsTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 /**
@@ -16,6 +17,30 @@ import org.junit.jupiter.params.provider.MethodSource;
  */
 @DisplayName("Utils:String")
 class StringUtilsTest {
+
+    @ParameterizedTest
+    @CsvSource(value = {"-1:false", "0:true", "a:false", "3 3:false"}, delimiterString = ":")
+    @DisplayName("유효한 숫자 여부 판별")
+    public void isNumberTest(String given, boolean expected) {
+        // When & Then
+        assertThat(StringUtils.isValidNumber(given)).isEqualTo(expected);
+    }
+
+    @ParameterizedTest
+    @MethodSource
+    @DisplayName("빈값 null 여부 판별")
+    public void isEmptyTest(String given, boolean expected) {
+        // When & Then
+        assertThat(StringUtils.isEmpty(given)).isEqualTo(expected);
+    }
+
+    public static Stream isEmptyTest() {
+        return Stream.of(
+            Arguments.of("", true),
+            Arguments.of(null, true),
+            Arguments.of(" ", true)
+        );
+    }
 
     @ParameterizedTest
     @MethodSource
